tests, partial.
