+++ apps/tests/curl_reconstruction.cpp
@@ -14,12 +14,9 @@
 
 #include <unistd.h>
 
-#include "bases/bases.hpp"
-#include "methods/hho"
-#include "methods/implementation_hho/curl.hpp"
-#include "quadratures/quadratures.hpp"
-
-#include "core/loaders/loader.hpp"
+#include "diskpp/loaders/loader.hpp"
+#include "diskpp/methods/hho"
+#include "diskpp/methods/implementation_hho/curl.hpp"
 
 #include "common.hpp"
