+++ apps/tests/curl_reconstruction_pk.cpp
@@ -14,17 +14,15 @@
 
 #include <unistd.h>
 
-#include "bases/bases.hpp"
-#include "methods/hho"
-#include "methods/implementation_hho/curl.hpp"
-#include "quadratures/quadratures.hpp"
+#include "diskpp/loaders/loader.hpp"
+#include "diskpp/methods/hho"
+#include "diskpp/methods/implementation_hho/curl.hpp"
 
-#include "core/loaders/loader.hpp"
 
 #include "common.hpp"
 
 template<typename Mesh, bool mixed>
-struct test_functor_curl_stab
+struct test_functor_curl_reconstruction
 {
     typename Mesh::coordinate_type
     operator()(const Mesh& msh, size_t degree) const
@@ -40,7 +38,7 @@ struct test_functor_curl_stab
 };
 
 template<template<typename, size_t, typename> class Mesh, typename T, typename Storage, bool mixed>
-struct test_functor_curl_stab<Mesh<T,3,Storage>, mixed>
+struct test_functor_curl_reconstruction<Mesh<T,3,Storage>, mixed>
 {
     using mesh_type = Mesh<T,3,Storage>;
     /* Expect k+1 convergence (hho stabilization) */
@@ -56,59 +54,72 @@ struct test_functor_curl_stab<Mesh<T,3,Storage>, mixed>
 
         auto f = [](const point_type& pt) -> Matrix<T,3,1> {
             Matrix<T,3,1> ret;
-            //ret(0) = std::sin(M_PI*pt.x());
-            //ret(1) = std::sin(M_PI*pt.y());
-            //ret(2) = std::sin(M_PI*pt.z());
-            ret(0) = 0.0;
-            ret(1) = 0.0;
-            ret(2) = pt.x()*pt.y()*std::sin(M_PI*pt.x())*std::sin(M_PI*pt.y());
-            //ret(2) = pt.x()*pt.x(); /* Beware of NaNs in std::sqrt */
+            ret(0) = std::sin(M_PI*pt.y());
+            ret(1) = std::sin(M_PI*pt.z());
+            ret(2) = std::sin(M_PI*pt.x());
             return ret;
         };
 
-        size_t fd = degree+1;
-        size_t cd = degree+1;
+        auto sol = [](const point_type& pt) -> Matrix<T,3,1> {
+            Matrix<T,3,1> ret;
+            ret(0) = -M_PI*std::cos(M_PI*pt.z());
+            ret(1) = -M_PI*std::cos(M_PI*pt.x());
+            ret(2) = -M_PI*std::cos(M_PI*pt.y());
+            return ret;
+        };
+
+        size_t fd = degree;
+        size_t cd = mixed ? degree+1 : degree;
         size_t rd = degree;
 
         disk::hho_degree_info hdi(disk::priv::hdi_named_args{.rd = rd, .cd = cd, .fd = fd});
 
         scalar_type error = 0.0;
-        for (auto& cl : msh)
+        for (const auto& cl : msh)
         {
-            auto stab = disk::curl_hdg_stabilization_nedelec(msh, cl, hdi);
-            Matrix<scalar_type, Dynamic, 1> proj = disk::project_tangent_nedelec(msh, cl, hdi, f, 1);
+            auto CR = disk::curl_reconstruction_pk(msh, cl, hdi);
+            auto proj = disk::project_tangent(msh, cl, hdi, f, 1);
+            Matrix<T, Dynamic, 1> rf = CR.first * proj;
 
-            error += proj.dot(stab * proj);
+            auto rb = disk::make_vector_monomial_basis(msh, cl, rd);
+
+            auto qps = integrate(msh, cl, 2*std::max(rd,cd)+1);
+            for (const auto& qp : qps)
+            {
+                auto cphi = rb.eval_functions(qp.point());
+                Matrix<T,3,1> diff = disk::eval(rf, cphi) - sol(qp.point());
+                error += qp.weight() * diff.dot(diff);
+            }
+
         }
 
-        return std::sqrt( std::abs(error) );
+        return std::sqrt(error);
     }
 
     size_t
     expected_rate(size_t k)
     {
-        return k + (mixed ? 1 : 0);
+        return k + 1;
     }
 };
 
 template<typename Mesh>
-using test_functor_curl_stab_eo = test_functor_curl_stab<Mesh, false>;
+using test_functor_curl_reconstruction_eo = test_functor_curl_reconstruction<Mesh, false>;
 
 template<typename Mesh>
-using test_functor_curl_stab_mo = test_functor_curl_stab<Mesh, true>;
+using test_functor_curl_reconstruction_mo = test_functor_curl_reconstruction<Mesh, true>;
 
 int
 main(void)
 {
-    std::cout << red << "Test HHO curl stabilization operator" << std::endl;
+    std::cout << red << "Test HHO curl reconstruction operator" << std::endl;
     // face order: k, cell order: k
     std::cout << cyan << "Face order: k and Cell order: k" << std::endl;
-    tester<test_functor_curl_stab_eo> tstr1;
+    tester<test_functor_curl_reconstruction_eo> tstr1;
     tstr1.run();
 
-    // face order: k, cell order: k
     std::cout << cyan << "Face order: k and Cell order: k+1" << std::endl;
-    tester<test_functor_curl_stab_mo> tstr2;
+    tester<test_functor_curl_reconstruction_mo> tstr2;
     tstr2.run();
 
     return 0;
