+++ apps/tests/curl_stabilization_nedelec.cpp
@@ -14,12 +14,10 @@
 
 #include <unistd.h>
 
-#include "bases/bases.hpp"
-#include "methods/hho"
-#include "methods/implementation_hho/curl.hpp"
-#include "quadratures/quadratures.hpp"
+#include "diskpp/loaders/loader.hpp"
+#include "diskpp/methods/hho"
+#include "diskpp/methods/implementation_hho/curl.hpp"
 
-#include "core/loaders/loader.hpp"
 
 #include "common.hpp"
 
@@ -36,47 +34,6 @@ struct test_functor_curl_stab
     expected_rate(size_t k)
     {
         return k + 50;
-    }
-};
-
-template<template<typename, size_t, typename> class Mesh, typename T, typename Storage, bool mixed>
-struct test_functor_curl_stab<Mesh<T,2,Storage>, mixed>
-{
-    using mesh_type = Mesh<T,2,Storage>;
-    /* Expect k+1 convergence (hho stabilization) */
-    typename mesh_type::coordinate_type
-    operator()(const mesh_type& msh, size_t degree) const
-    {
-        typedef typename mesh_type::cell            cell_type;
-        typedef typename mesh_type::face            face_type;
-        typedef typename mesh_type::coordinate_type scalar_type;
-        typedef typename mesh_type::point_type      point_type;
-
-        typedef Matrix<scalar_type, mesh_type::dimension, 1> ret_type;
-
-        auto f = [](const point_type& pt) -> T {
-            return std::sin(M_PI*pt.x())*std::sin(M_PI*pt.y());
-        };
-
-        typename disk::hho_degree_info hdi(mixed ? degree+1 : degree, degree);
-
-        scalar_type error = 0.0;
-        for (auto& cl : msh)
-        {
-            auto stab = disk::curl_hdg_stabilization(msh, cl, hdi);
-
-            Matrix<scalar_type, Dynamic, 1> proj = disk::project_tangent(msh, cl, hdi, f);
-
-            error += proj.dot(stab * proj);
-        }
-
-        return std::sqrt(error);
-    }
-
-    size_t
-    expected_rate(size_t k)
-    {
-        return k + (mixed ? 1 : 0);
     }
 };
 
@@ -107,16 +64,17 @@ struct test_functor_curl_stab<Mesh<T,3,Storage>, mixed>
             return ret;
         };
 
-        disk::hho_degree_info hdi( mixed ? degree+1 : degree, degree );
+        size_t fd = degree+1;
+        size_t cd = degree+1;
+        size_t rd = degree;
+
+        disk::hho_degree_info hdi(disk::priv::hdi_named_args{.rd = rd, .cd = cd, .fd = fd});
 
         scalar_type error = 0.0;
         for (auto& cl : msh)
         {
-            //auto CR = disk::curl_reconstruction(msh, cl, hdi);
-            //auto stab = disk::curl_hho_stabilization(msh, cl, CR.first, hdi);
-
-            auto stab = disk::curl_hdg_stabilization(msh, cl, hdi);
-            Matrix<scalar_type, Dynamic, 1> proj = disk::project_tangent(msh, cl, hdi, f, 1);
+            auto stab = disk::curl_hdg_stabilization_nedelec(msh, cl, hdi);
+            Matrix<scalar_type, Dynamic, 1> proj = disk::project_tangent_nedelec(msh, cl, hdi, f, 1);
 
             error += proj.dot(stab * proj);
         }
