+++ apps/tests/divergence_reconstruction.cpp
@@ -24,23 +24,17 @@
  * DOI: 10.1016/j.cam.2017.09.017
  */
 
-#include <iostream>
 #include <iomanip>
+#include <iostream>
 #include <regex>
 
-#include "contrib/colormanip.h"
 #include <unistd.h>
 
-#include <xmmintrin.h>
-//#define EIGEN_USE_MKL_ALL
-#include "bases/bases.hpp"
-#include "quadratures/quadratures.hpp"
-#include "methods/hho"
+#include "diskpp/loaders/loader.hpp"
+#include "diskpp/methods/hho"
 
-#include "core/loaders/loader.hpp"
 
 #include "common.hpp"
-
 
 template<typename Mesh>
 struct test_functor_equal_order
@@ -49,14 +43,16 @@ struct test_functor_equal_order
     typename Mesh::coordinate_type
     operator()(const Mesh& msh, size_t degree) const
     {
-        typedef Mesh mesh_type;
-        typedef typename mesh_type::cell        cell_type;
-        typedef typename mesh_type::face        face_type;
+        typedef Mesh                                mesh_type;
+        typedef typename mesh_type::cell            cell_type;
+        typedef typename mesh_type::face            face_type;
         typedef typename mesh_type::coordinate_type scalar_type;
-        typedef typename mesh_type::point_type  point_type;
+        typedef typename mesh_type::point_type      point_type;
 
+        typedef Matrix<scalar_type, Mesh::dimension, 1> ret_type;
 
-        auto f = make_scalar_testing_data(msh);
+        auto f    = make_vector_testing_data(msh);
+        auto divf = make_vector_testing_data_div(msh);
 
         typename disk::hho_degree_info hdi(degree);
 
@@ -64,24 +60,18 @@ struct test_functor_equal_order
         for (auto& cl : msh)
         {
             Matrix<scalar_type, Dynamic, 1> proj = disk::project_function(msh, cl, hdi, f, 2);
-            auto gr = disk::make_scalar_hho_laplacian(msh, cl, hdi);
+            const auto                      dr   = make_hho_divergence_reconstruction(msh, cl, hdi);
 
-            size_t rec_size = disk::scalar_basis_size(hdi.reconstruction_degree(), Mesh::dimension);
+            Matrix<scalar_type, Dynamic, 1> div = dr.first * proj;
 
-            Matrix<scalar_type, Dynamic, 1> reconstr = Matrix<scalar_type, Dynamic, 1>::Zero(rec_size);
-            reconstr.tail(rec_size-1) = gr.first * proj;
-            reconstr(0) = proj(0);
-
-            auto cb = disk::make_scalar_monomial_basis(msh, cl, hdi.reconstruction_degree());
+            const auto                            cb   = disk::make_scalar_monomial_basis(msh, cl, hdi.face_degree());
             Matrix<scalar_type, Dynamic, Dynamic> mass = disk::make_mass_matrix(msh, cl, cb);
-            Matrix<scalar_type, Dynamic, 1> rhs = disk::make_rhs(msh, cl, cb, f);
-            Matrix<scalar_type, Dynamic, 1> exp_reconstr = mass.llt().solve(rhs);
+            Matrix<scalar_type, Dynamic, 1>       rhs  = disk::make_rhs(msh, cl, cb, divf);
+            Matrix<scalar_type, Dynamic, 1>       exp_div = mass.llt().solve(rhs);
 
-            Matrix<scalar_type, Dynamic, 1> diff = reconstr - exp_reconstr;
+            Matrix<scalar_type, Dynamic, 1> diff = div - exp_div;
 
-            Matrix<scalar_type, Dynamic, Dynamic> stiffness = disk::make_stiffness_matrix(msh, cl, cb);
-
-            error += diff.dot(stiffness*diff);
+            error += diff.dot(mass * diff);
         }
 
         return std::sqrt(error);
@@ -90,7 +80,7 @@ struct test_functor_equal_order
     size_t
     expected_rate(size_t k)
     {
-        return k+1;
+        return k + 1;
     }
 };
 
@@ -107,32 +97,29 @@ struct test_functor_mixed_order1
         typedef typename mesh_type::coordinate_type scalar_type;
         typedef typename mesh_type::point_type      point_type;
 
-        auto f = make_scalar_testing_data(msh);
+        typedef Matrix<scalar_type, Mesh::dimension, 1> ret_type;
 
-        typename disk::hho_degree_info hdi(degree+1, degree);
+        auto f    = make_vector_testing_data(msh);
+        auto divf = make_vector_testing_data_div(msh);
+
+        typename disk::hho_degree_info hdi(degree + 1, degree);
 
         scalar_type error = 0.0;
         for (auto& cl : msh)
         {
             Matrix<scalar_type, Dynamic, 1> proj = disk::project_function(msh, cl, hdi, f, 2);
-            auto                            gr   = disk::make_scalar_hho_laplacian(msh, cl, hdi);
+            const auto                      dr   = make_hho_divergence_reconstruction(msh, cl, hdi);
 
-            size_t rec_size = disk::scalar_basis_size(hdi.reconstruction_degree(), Mesh::dimension);
+            Matrix<scalar_type, Dynamic, 1> div = dr.first * proj;
 
-            Matrix<scalar_type, Dynamic, 1> reconstr = Matrix<scalar_type, Dynamic, 1>::Zero(rec_size);
-            reconstr.tail(rec_size - 1)              = gr.first * proj;
-            reconstr(0)                              = proj(0);
+            const auto                            cb   = disk::make_scalar_monomial_basis(msh, cl, hdi.face_degree());
+            Matrix<scalar_type, Dynamic, Dynamic> mass = disk::make_mass_matrix(msh, cl, cb);
+            Matrix<scalar_type, Dynamic, 1>       rhs  = disk::make_rhs(msh, cl, cb, divf);
+            Matrix<scalar_type, Dynamic, 1>       exp_div = mass.llt().solve(rhs);
 
-            auto cb = disk::make_scalar_monomial_basis(msh, cl, hdi.reconstruction_degree());
-            Matrix<scalar_type, Dynamic, Dynamic> mass         = disk::make_mass_matrix(msh, cl, cb);
-            Matrix<scalar_type, Dynamic, 1>       rhs          = disk::make_rhs(msh, cl, cb, f);
-            Matrix<scalar_type, Dynamic, 1>       exp_reconstr = mass.llt().solve(rhs);
+            Matrix<scalar_type, Dynamic, 1> diff = div - exp_div;
 
-            Matrix<scalar_type, Dynamic, 1> diff = reconstr - exp_reconstr;
-
-            Matrix<scalar_type, Dynamic, Dynamic> stiffness = disk::make_stiffness_matrix(msh, cl, cb);
-
-            error += diff.dot(stiffness * diff);
+            error += diff.dot(mass * diff);
         }
 
         return std::sqrt(error);
@@ -158,32 +145,29 @@ struct test_functor_mixed_order2
         typedef typename mesh_type::coordinate_type scalar_type;
         typedef typename mesh_type::point_type      point_type;
 
-        auto f = make_scalar_testing_data(msh);
+        typedef Matrix<scalar_type, Mesh::dimension, 1> ret_type;
 
-        typename disk::hho_degree_info hdi(degree-1, degree);
+        auto f    = make_vector_testing_data(msh);
+        auto divf = make_vector_testing_data_div(msh);
+
+        typename disk::hho_degree_info hdi(degree - 1, degree);
 
         scalar_type error = 0.0;
         for (auto& cl : msh)
         {
             Matrix<scalar_type, Dynamic, 1> proj = disk::project_function(msh, cl, hdi, f, 2);
-            auto                            gr   = disk::make_scalar_hho_laplacian(msh, cl, hdi);
+            const auto                      dr   = make_hho_divergence_reconstruction(msh, cl, hdi);
 
-            size_t rec_size = disk::scalar_basis_size(hdi.reconstruction_degree(), Mesh::dimension);
+            Matrix<scalar_type, Dynamic, 1> div = dr.first * proj;
 
-            Matrix<scalar_type, Dynamic, 1> reconstr = Matrix<scalar_type, Dynamic, 1>::Zero(rec_size);
-            reconstr.tail(rec_size - 1)              = gr.first * proj;
-            reconstr(0)                              = proj(0);
+            const auto                            cb   = disk::make_scalar_monomial_basis(msh, cl, hdi.face_degree());
+            Matrix<scalar_type, Dynamic, Dynamic> mass = disk::make_mass_matrix(msh, cl, cb);
+            Matrix<scalar_type, Dynamic, 1>       rhs  = disk::make_rhs(msh, cl, cb, divf);
+            Matrix<scalar_type, Dynamic, 1>       exp_div = mass.llt().solve(rhs);
 
-            auto cb = disk::make_scalar_monomial_basis(msh, cl, hdi.reconstruction_degree());
-            Matrix<scalar_type, Dynamic, Dynamic> mass         = disk::make_mass_matrix(msh, cl, cb);
-            Matrix<scalar_type, Dynamic, 1>       rhs          = disk::make_rhs(msh, cl, cb, f);
-            Matrix<scalar_type, Dynamic, 1>       exp_reconstr = mass.llt().solve(rhs);
+            Matrix<scalar_type, Dynamic, 1> diff = div - exp_div;
 
-            Matrix<scalar_type, Dynamic, 1> diff = reconstr - exp_reconstr;
-
-            Matrix<scalar_type, Dynamic, Dynamic> stiffness = disk::make_stiffness_matrix(msh, cl, cb);
-
-            error += diff.dot(stiffness * diff);
+            error += diff.dot(mass * diff);
         }
 
         return std::sqrt(error);
@@ -196,7 +180,8 @@ struct test_functor_mixed_order2
     }
 };
 
-int main(void)
+int
+main(void)
 {
     // face order: k, cell order: k
     std::cout << blue << "Face order: k and Cell order: k" << std::endl;
@@ -209,6 +194,6 @@ int main(void)
     // face order: k, cell order: k-1
     std::cout << blue << "Face order: k and Cell order: k-1" << std::endl;
     tester<test_functor_mixed_order2> tstr3;
-    tstr3.run(1,3);
+    tstr3.run(1, 3);
     return 0;
 }
