+++ apps/tests/divide_polygons.cpp
@@ -29,11 +29,11 @@
 #include <regex>
 #include <unistd.h>
 
-#include "loaders/loader.hpp"
+#include "diskpp/loaders/loader.hpp"
 
-#include "output/gmshConvertMesh.hpp"
-#include "output/gmshDisk.hpp"
-#include "output/postMesh.hpp"
+#include "diskpp/output/gmshConvertMesh.hpp"
+#include "diskpp/output/gmshDisk.hpp"
+#include "diskpp/output/postMesh.hpp"
 
 template<typename MeshType>
 void split_2D_polygons(MeshType& msh)
