+++ apps/tests/raw_quadratures.cpp
@@ -27,9 +27,9 @@
 
 #include <xmmintrin.h>
 
-#include "bases/bases.hpp"
-#include "quadratures/quadratures.hpp"
-#include "methods/hho"
+#include "diskpp/bases/bases.hpp"
+#include "diskpp/quadratures/quadratures.hpp"
+#include "diskpp/methods/hho"
 
 #include "common.hpp"
