+++ apps/tests/scalar_reconstruction.cpp
@@ -24,19 +24,19 @@
  * DOI: 10.1016/j.cam.2017.09.017
  */
 
-#include <iomanip>
 #include <iostream>
+#include <iomanip>
 #include <regex>
 
 #include <unistd.h>
 
-#include "bases/bases.hpp"
-#include "methods/hho"
-#include "quadratures/quadratures.hpp"
-
-#include "core/loaders/loader.hpp"
+#include <xmmintrin.h>
+#include "diskpp/loaders/loader.hpp"
+#include "diskpp/methods/hho"
+#include "diskpp/common/colormanip.h"
 
 #include "common.hpp"
+
 
 template<typename Mesh>
 struct test_functor_equal_order
@@ -45,42 +45,48 @@ struct test_functor_equal_order
     typename Mesh::coordinate_type
     operator()(const Mesh& msh, size_t degree) const
     {
-        typedef Mesh                                mesh_type;
-        typedef typename mesh_type::cell            cell_type;
-        typedef typename mesh_type::face            face_type;
+        typedef Mesh mesh_type;
+        typedef typename mesh_type::cell        cell_type;
+        typedef typename mesh_type::face        face_type;
         typedef typename mesh_type::coordinate_type scalar_type;
-        typedef typename mesh_type::point_type      point_type;
+        typedef typename mesh_type::point_type  point_type;
 
-        typedef Matrix<scalar_type, Mesh::dimension, 1> ret_type;
 
-        auto f  = make_scalar_testing_data(msh);
-        auto gf = make_scalar_testing_data_grad(msh);
+        auto f = make_scalar_testing_data(msh);
 
-        typename disk::hho_degree_info hdi(degree, degree, degree+1);
+        typename disk::hho_degree_info hdi(degree);
 
         scalar_type error = 0.0;
         for (auto& cl : msh)
         {
             Matrix<scalar_type, Dynamic, 1> proj = disk::project_function(msh, cl, hdi, f, 2);
-            const auto                      gr   = disk::make_vector_hho_gradrec(msh, cl, hdi);
+            auto gr = disk::make_scalar_hho_laplacian(msh, cl, hdi);
 
-            Matrix<scalar_type, Dynamic, 1> gradc = gr.first * proj;
-            Matrix<scalar_type, Dynamic, 1> grad = disk::project_function(msh, cl, hdi.grad_degree(), gf, 2);
-            Matrix<scalar_type, Dynamic, 1> diff = grad - gradc;
+            size_t rec_size = disk::scalar_basis_size(hdi.reconstruction_degree(), Mesh::dimension);
 
-            const auto                            gb   = disk::make_vector_monomial_basis(msh, cl, hdi.grad_degree());
-            Matrix<scalar_type, Dynamic, Dynamic> mass = disk::make_mass_matrix(msh, cl, gb);
+            Matrix<scalar_type, Dynamic, 1> reconstr = Matrix<scalar_type, Dynamic, 1>::Zero(rec_size);
+            reconstr.tail(rec_size-1) = gr.first * proj;
+            reconstr(0) = proj(0);
 
-            error += diff.dot(mass * diff);
+            auto cb = disk::make_scalar_monomial_basis(msh, cl, hdi.reconstruction_degree());
+            Matrix<scalar_type, Dynamic, Dynamic> mass = disk::make_mass_matrix(msh, cl, cb);
+            Matrix<scalar_type, Dynamic, 1> rhs = disk::make_rhs(msh, cl, cb, f);
+            Matrix<scalar_type, Dynamic, 1> exp_reconstr = mass.llt().solve(rhs);
+
+            Matrix<scalar_type, Dynamic, 1> diff = reconstr - exp_reconstr;
+
+            Matrix<scalar_type, Dynamic, Dynamic> stiffness = disk::make_stiffness_matrix(msh, cl, cb);
+
+            error += diff.dot(stiffness*diff);
         }
 
         return std::sqrt(error);
     }
 
-    size_t
+    double
     expected_rate(size_t k)
     {
-        return k + 1;
+        return k+1;
     }
 };
 
@@ -97,26 +103,23 @@ struct test_functor_mixed_order1
         typedef typename mesh_type::coordinate_type scalar_type;
         typedef typename mesh_type::point_type      point_type;
 
-        typedef Matrix<scalar_type, Mesh::dimension, 1> ret_type;
+        auto f = make_scalar_testing_data(msh);
 
-        auto f = make_vector_testing_data(msh);
-
-        typename disk::hho_degree_info hdi(degree + 1, degree);
+        typename disk::hho_degree_info hdi(degree+1, degree);
 
         scalar_type error = 0.0;
         for (auto& cl : msh)
         {
             Matrix<scalar_type, Dynamic, 1> proj = disk::project_function(msh, cl, hdi, f, 2);
-            auto                            gr   = disk::make_vector_hho_laplacian(msh, cl, hdi);
+            auto                            gr   = disk::make_scalar_hho_laplacian(msh, cl, hdi);
 
-            size_t rec_size = disk::vector_basis_size(hdi.reconstruction_degree(), Mesh::dimension, Mesh::dimension);
+            size_t rec_size = disk::scalar_basis_size(hdi.reconstruction_degree(), Mesh::dimension);
 
-            Matrix<scalar_type, Dynamic, 1> reconstr  = Matrix<scalar_type, Dynamic, 1>::Zero(rec_size);
-            reconstr.tail(rec_size - Mesh::dimension) = gr.first * proj;
+            Matrix<scalar_type, Dynamic, 1> reconstr = Matrix<scalar_type, Dynamic, 1>::Zero(rec_size);
+            reconstr.tail(rec_size - 1)              = gr.first * proj;
+            reconstr(0)                              = proj(0);
 
-            reconstr.head(Mesh::dimension) = proj.head(Mesh::dimension);
-
-            auto cb = disk::make_vector_monomial_basis(msh, cl, hdi.reconstruction_degree());
+            auto cb = disk::make_scalar_monomial_basis(msh, cl, hdi.reconstruction_degree());
             Matrix<scalar_type, Dynamic, Dynamic> mass         = disk::make_mass_matrix(msh, cl, cb);
             Matrix<scalar_type, Dynamic, 1>       rhs          = disk::make_rhs(msh, cl, cb, f);
             Matrix<scalar_type, Dynamic, 1>       exp_reconstr = mass.llt().solve(rhs);
@@ -131,7 +134,7 @@ struct test_functor_mixed_order1
         return std::sqrt(error);
     }
 
-    size_t
+    double
     expected_rate(size_t k)
     {
         return k + 1;
@@ -151,26 +154,23 @@ struct test_functor_mixed_order2
         typedef typename mesh_type::coordinate_type scalar_type;
         typedef typename mesh_type::point_type      point_type;
 
-        typedef Matrix<scalar_type, Mesh::dimension, 1> ret_type;
+        auto f = make_scalar_testing_data(msh);
 
-        auto f = make_vector_testing_data(msh);
-
-        typename disk::hho_degree_info hdi(degree - 1, degree);
+        typename disk::hho_degree_info hdi(degree-1, degree);
 
         scalar_type error = 0.0;
         for (auto& cl : msh)
         {
             Matrix<scalar_type, Dynamic, 1> proj = disk::project_function(msh, cl, hdi, f, 2);
-            auto                            gr   = disk::make_vector_hho_laplacian(msh, cl, hdi);
+            auto                            gr   = disk::make_scalar_hho_laplacian(msh, cl, hdi);
 
-            size_t rec_size = disk::vector_basis_size(hdi.reconstruction_degree(), Mesh::dimension, Mesh::dimension);
+            size_t rec_size = disk::scalar_basis_size(hdi.reconstruction_degree(), Mesh::dimension);
 
-            Matrix<scalar_type, Dynamic, 1> reconstr  = Matrix<scalar_type, Dynamic, 1>::Zero(rec_size);
-            reconstr.tail(rec_size - Mesh::dimension) = gr.first * proj;
+            Matrix<scalar_type, Dynamic, 1> reconstr = Matrix<scalar_type, Dynamic, 1>::Zero(rec_size);
+            reconstr.tail(rec_size - 1)              = gr.first * proj;
+            reconstr(0)                              = proj(0);
 
-            reconstr.head(Mesh::dimension) = proj.head(Mesh::dimension);
-
-            auto cb = disk::make_vector_monomial_basis(msh, cl, hdi.reconstruction_degree());
+            auto cb = disk::make_scalar_monomial_basis(msh, cl, hdi.reconstruction_degree());
             Matrix<scalar_type, Dynamic, Dynamic> mass         = disk::make_mass_matrix(msh, cl, cb);
             Matrix<scalar_type, Dynamic, 1>       rhs          = disk::make_rhs(msh, cl, cb, f);
             Matrix<scalar_type, Dynamic, 1>       exp_reconstr = mass.llt().solve(rhs);
@@ -185,15 +185,14 @@ struct test_functor_mixed_order2
         return std::sqrt(error);
     }
 
-    size_t
+    double 
     expected_rate(size_t k)
     {
         return k + 1;
     }
 };
 
-int
-main(void)
+int main(void)
 {
     // face order: k, cell order: k
     std::cout << blue << "Face order: k and Cell order: k" << std::endl;
@@ -206,6 +205,6 @@ main(void)
     // face order: k, cell order: k-1
     std::cout << blue << "Face order: k and Cell order: k-1" << std::endl;
     tester<test_functor_mixed_order2> tstr3;
-    tstr3.run(1, 3);
+    tstr3.run(1,3);
     return 0;
 }
