+++ apps/tests/vector_gradient.cpp
@@ -24,18 +24,16 @@
  * DOI: 10.1016/j.cam.2017.09.017
  */
 
-
-#include <iostream>
 #include <iomanip>
+#include <iostream>
 #include <regex>
 
 #include <unistd.h>
 
-#include "bases/bases.hpp"
-#include "quadratures/quadratures.hpp"
-#include "methods/hho"
-
-#include "core/loaders/loader.hpp"
+#include "diskpp/bases/bases.hpp"
+#include "diskpp/methods/hho"
+#include "diskpp/quadratures/quadratures.hpp"
+#include "diskpp/loaders/loader.hpp"
 
 #include "common.hpp"
 
@@ -46,41 +44,33 @@ struct test_functor_equal_order
     typename Mesh::coordinate_type
     operator()(const Mesh& msh, size_t degree) const
     {
-        typedef Mesh mesh_type;
-        typedef typename mesh_type::cell        cell_type;
-        typedef typename mesh_type::face        face_type;
+        typedef Mesh                                mesh_type;
+        typedef typename mesh_type::cell            cell_type;
+        typedef typename mesh_type::face            face_type;
         typedef typename mesh_type::coordinate_type scalar_type;
-        typedef typename mesh_type::point_type  point_type;
+        typedef typename mesh_type::point_type      point_type;
 
         typedef Matrix<scalar_type, Mesh::dimension, 1> ret_type;
 
-        auto f = make_vector_testing_data(msh);
+        auto f  = make_scalar_testing_data(msh);
+        auto gf = make_scalar_testing_data_grad(msh);
 
-        typename disk::hho_degree_info hdi(degree);
+        typename disk::hho_degree_info hdi(degree, degree, degree+1);
 
         scalar_type error = 0.0;
         for (auto& cl : msh)
         {
             Matrix<scalar_type, Dynamic, 1> proj = disk::project_function(msh, cl, hdi, f, 2);
-            auto gr = disk::make_vector_hho_laplacian(msh, cl, hdi);
+            const auto                      gr   = disk::make_vector_hho_gradrec(msh, cl, hdi);
 
-            size_t rec_size = disk::vector_basis_size(hdi.reconstruction_degree(), Mesh::dimension, Mesh::dimension);
+            Matrix<scalar_type, Dynamic, 1> gradc = gr.first * proj;
+            Matrix<scalar_type, Dynamic, 1> grad = disk::project_function(msh, cl, hdi.grad_degree(), gf, 2);
+            Matrix<scalar_type, Dynamic, 1> diff = grad - gradc;
 
-            Matrix<scalar_type, Dynamic, 1> reconstr = Matrix<scalar_type, Dynamic, 1>::Zero(rec_size);
-            reconstr.tail(rec_size-Mesh::dimension) = gr.first * proj;
+            const auto                            gb   = disk::make_vector_monomial_basis(msh, cl, hdi.grad_degree());
+            Matrix<scalar_type, Dynamic, Dynamic> mass = disk::make_mass_matrix(msh, cl, gb);
 
-            reconstr.head(Mesh::dimension) = proj.head(Mesh::dimension);
-
-            auto cb = disk::make_vector_monomial_basis(msh, cl, hdi.reconstruction_degree());
-            Matrix<scalar_type, Dynamic, Dynamic> mass = disk::make_mass_matrix(msh, cl, cb);
-            Matrix<scalar_type, Dynamic, 1> rhs = disk::make_rhs(msh, cl, cb, f);
-            Matrix<scalar_type, Dynamic, 1> exp_reconstr = mass.llt().solve(rhs);
-
-            Matrix<scalar_type, Dynamic, 1> diff = reconstr - exp_reconstr;
-
-            Matrix<scalar_type, Dynamic, Dynamic> stiffness = disk::make_stiffness_matrix(msh, cl, cb);
-
-            error += diff.dot(stiffness*diff);
+            error += diff.dot(mass * diff);
         }
 
         return std::sqrt(error);
@@ -89,7 +79,7 @@ struct test_functor_equal_order
     size_t
     expected_rate(size_t k)
     {
-        return k+1;
+        return k + 1;
     }
 };
 
@@ -110,7 +100,7 @@ struct test_functor_mixed_order1
 
         auto f = make_vector_testing_data(msh);
 
-        typename disk::hho_degree_info hdi(degree+1, degree);
+        typename disk::hho_degree_info hdi(degree + 1, degree);
 
         scalar_type error = 0.0;
         for (auto& cl : msh)
@@ -164,7 +154,7 @@ struct test_functor_mixed_order2
 
         auto f = make_vector_testing_data(msh);
 
-        typename disk::hho_degree_info hdi(degree-1, degree);
+        typename disk::hho_degree_info hdi(degree - 1, degree);
 
         scalar_type error = 0.0;
         for (auto& cl : msh)
@@ -201,7 +191,8 @@ struct test_functor_mixed_order2
     }
 };
 
-int main(void)
+int
+main(void)
 {
     // face order: k, cell order: k
     std::cout << blue << "Face order: k and Cell order: k" << std::endl;
