+++ apps/tests/vector_symmetric_stabilization.cpp
@@ -30,11 +30,8 @@
 
 #include <unistd.h>
 
-#include "bases/bases.hpp"
-#include "methods/hho"
-#include "quadratures/quadratures.hpp"
-
-#include "core/loaders/loader.hpp"
+#include "diskpp/loaders/loader.hpp"
+#include "diskpp/methods/hho"
 
 #include "common.hpp"
