+++ libdiskpp/include/diskpp/common/timecounter.hpp
@@ -1,0 +1,77 @@
+/*
+ * This source file is part of EMT, the ElectroMagneticTool.
+ *
+ * Copyright (C) 2013-2015, Matteo Cicuttin - matteo.cicuttin@uniud.it
+ * Department of Electrical Engineering, University of Udine
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *     * Redistributions of source code must retain the above copyright
+ *       notice, this list of conditions and the following disclaimer.
+ *     * Redistributions in binary form must reproduce the above copyright
+ *       notice, this list of conditions and the following disclaimer in the
+ *       documentation and/or other materials provided with the distribution.
+ *     * Neither the name of the University of Udine nor the
+ *       names of its contributors may be used to endorse or promote products
+ *       derived from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE AUTHOR(s) ``AS IS'' AND ANY
+ * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE AUTHOR(s) BE LIABLE FOR ANY
+ * DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#pragma once
+
+#include <chrono>
+#include <iostream>
+#include <string>
+#include <map>
+
+class timecounter
+{
+    std::chrono::time_point<std::chrono::steady_clock>    begin, end;
+    bool running;
+
+public:
+    timecounter() : running(false)
+    {}
+
+    void tic()
+    {
+        begin = std::chrono::steady_clock::now();
+        running = true;
+    }
+
+    double toc()
+    {
+        end = std::chrono::steady_clock::now();
+        running = false;
+        std::chrono::duration<double> diff = end-begin;
+        return diff.count();
+    }
+
+    double elapsed() const
+    {
+        if (not running)
+            return std::chrono::duration<double>(end-begin).count();
+        
+        std::chrono::duration<double> diff =
+            std::chrono::steady_clock::now()-begin;
+        return diff.count();
+    }
+};
+
+std::ostream&
+operator<<(std::ostream& os, const timecounter& tc)
+{
+    os << tc.elapsed();
+    return os;
+}
