+++ libdiskpp/include/diskpp/mechanics/behaviors/maths_tensor.hpp
@@ -25,7 +25,7 @@
 
 #pragma once
 
-#include "common/eigen.hpp"
+#include "diskpp/common/eigen.hpp"
 
 namespace disk
 {
