+++ libdiskpp/include/diskpp/mechanics/deformation_tensors.hpp
@@ -27,7 +27,7 @@
 
 #pragma once
 
-#include "common/eigen.hpp"
+#include "diskpp/common/eigen.hpp"
 
 namespace disk
 {
